add auth

+++ memberManagement_service/src/main/java/example/memberManagement_service.java
@@ -2,10 +2,10 @@ package example;
 
 import java.util.Collections;
 import java.util.List;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,8 +15,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.reactive.function.client.WebClient;
-
+import org.springframework.web.bind.annotation.RequestHeader;
 import reactor.core.publisher.Mono;
 
 @RestController
@@ -55,48 +56,71 @@ public class memberManagement_service {
 
     // Get user by id
     @GetMapping("/members/{id}")
-    List<Member> getMember(@PathVariable String id) {
-        List<Member> member = database.get()
+    ResponseEntity<Object> getMember(@RequestHeader("Authorization") String token, @PathVariable String id) {
+        Boolean valid = database.get()
+                .uri("/auth/validate")
+                .header("Authorization", token)
+                .retrieve()
+                .bodyToMono(Boolean.class).block();
+        if (!valid)
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
+        Member member = database.get()
                 .uri("/members/" + id)
                 .retrieve()
-                .bodyToFlux(Member.class)
-                .collectList()
-                .block();
-        return member;
+                .bodyToMono(Member.class).block();
+        return new ResponseEntity<>(member, HttpStatus.OK);
     }
 
     // Create Member
     @PostMapping("/members")
     Member createMember(@RequestBody Member m) {
-        database.post()
+        Member out = database.post()
                 .uri("/members")
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(Mono.just(m), Member.class)
                 .retrieve()
                 .bodyToMono(Member.class).block();
-        return m;
+        return out;
     }
 
     // Update Member
     @PutMapping("/members/{id}")
-    Member updateMember(@PathVariable String id, @RequestBody Member m) {
+    ResponseEntity<Object> updateMember(@RequestHeader("Authorization") String token, @PathVariable String id,
+            @RequestBody Member m) {
+        boolean valid = database.get()
+                .uri("/auth/validate")
+                .header("Authorization", token)
+                .retrieve()
+                .bodyToMono(Boolean.class).block();
+
+        if (!valid)
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         database.put()
                 .uri("/members/" + id)
                 .contentType(MediaType.APPLICATION_JSON)
                 .body(Mono.just(m), Member.class)
                 .retrieve()
                 .bodyToMono(Member.class).block();
-        return m;
+        return new ResponseEntity<>(m, HttpStatus.OK);
 
     }
 
     // Delete Member
     @DeleteMapping("/members/{id}")
-    void deleteMember(@PathVariable String id) {
+    ResponseEntity<Object> deleteMember(@RequestHeader("Authorization") String token, @PathVariable String id) {
+        boolean valid = database.get()
+                .uri("/auth/validate")
+                .header("Authorization", token)
+                .retrieve()
+                .bodyToMono(Boolean.class).block();
+
+        if (!valid)
+            return new ResponseEntity<>("Unauthorized", HttpStatus.UNAUTHORIZED);
         database.delete()
                 .uri("members/" + id)
                 .retrieve()
                 .bodyToMono(Void.class).block();
+        return new ResponseEntity<>("Deleted", HttpStatus.OK);
 
     }
